requirements.txt
tensorflow>=2.15.0
keras==3.10.0
flask
pillow
numpy
opencv-python
matplotlib